Add documentation. Rename model gallery to model gallery and clocks to models
