+++ biolearn/data_library.py
@@ -1,7 +1,7 @@
 import yaml
 import pandas as pd
 import re
-from biolearn.util import cached_dowload, get_data_file
+from biolearn.util import cached_download, get_data_file
 
 
 def parse_after_colon(s):
@@ -30,12 +30,28 @@ def extract_numeric(s):
 
 
 class GeoData:
-    def __init__(self, metadata, dnam):
-        # Metadata should have rows being samples and columns being data fields
-        self.metadata = metadata
-        # Methylation data should have columns as samples and rows as methylation sites
-        self.dnam = dnam
+    """
+    Represents genomic data with a focus on metadata and methylation data.
 
+    GeoData facilitates the organization and access to metadata and methylation data.
+
+    Attributes:
+        metadata (DataFrame): A pandas DataFrame where rows represent different samples
+                              and columns represent different data fields.
+        dnam (DataFrame): A pandas DataFrame where columns represent different samples
+                          and rows represent different methylation sites.
+    """
+
+    def __init__(self, metadata, dnam):
+        """
+        Initializes the GeoData instance.
+
+        Args:
+            metadata (DataFrame): Metadata associated with genomic samples.
+            dnam (DataFrame): Methylation data associated with genomic samples.
+        """
+        self.metadata = metadata
+        self.dnam = dnam
 
 class GeoMatrixParser:
     parsers = {
@@ -84,6 +100,26 @@ class GeoMatrixParser:
 
 
 class DataSource:
+    """
+    Represents a single data source in the DataLibrary.
+
+    This class encapsulates the details of a data source, including its identifier, path, 
+    parser, and optional metadata such as title, summary, format, and organism. It also 
+    provides functionality to load and parse the data from the source.
+
+    Attributes:
+        id (str): Identifier of the data source.
+        path (str): Path where the data source is located.
+        parser (object): Parser object responsible for parsing the data source.
+        title (str, optional): Title of the data source.
+        summary (str, optional): Summary of the data source.
+        format (str, optional): Format of the data source.
+        organism (str, optional): Organism related to the data source.
+
+    Raises:
+        ValueError: If any of the required fields ('id', 'path', 'parser') are missing
+                    during initialization.
+    """
     REQUIRED_FIELDS = {
         "id": "'id' key is missing in item",
         "path": "'path' key is missing in item",
@@ -91,6 +127,15 @@ class DataSource:
     }
 
     def __init__(self, data):
+        """
+        Initializes the DataSource instance. Intended to be initialized with data from a library YAML file.
+
+        Args:
+            data (dict): A dictionary containing the data source's properties.
+
+        Raises:
+            ValueError: If any of the required fields ('id', 'path', 'parser') are missing.
+        """
         for field, error_message in self.REQUIRED_FIELDS.items():
             setattr(self, field, data.get(field))
             if getattr(self, field) is None:
@@ -103,6 +148,26 @@ class DataSource:
 
         self.parser = self._create_parser(data["parser"])
 
+    def load(self):
+        """
+        Loads the data from the source.
+
+        Returns:
+            DNA methylation and metadata
+        """
+        file_path = cached_download(self.path)
+        return self.parser.parse(file_path)
+
+    def __repr__(self):
+        """
+        Represents the DataSource instance as a string.
+
+        Returns:
+            str: A string representation of the DataSource instance.
+        """
+        return f"DataSource(ID: {self.id}, Title: {self.title})"
+
+
     def _create_parser(self, parser_data):
         parser_type = parser_data.get("type")
         if parser_type is None:
@@ -111,13 +176,6 @@ class DataSource:
             return GeoMatrixParser(parser_data)
         # Add more parsers as needed
         raise ValueError(f"Unknown parser type: {parser_type}")
-
-    def load(self):
-        file_path = cached_dowload(self.path)
-        return self.parser.parse(file_path)
-
-    def __repr__(self):
-        return f"DataSource(ID: {self.id}\nTitle: {self.title})"
 
 
 def parse_library_file(library_file):
@@ -132,24 +190,69 @@ def parse_library_file(library_file):
 
 
 class DataLibrary:
+    """
+    Manages a collection of data sources for biomarkers research.
+
+    The DataLibrary class is responsible for loading, storing, and retrieving data sources
+    from a specified library file. It offers functionality to look up data sources based on
+    specific criteria such as organism and format, and to retrieve a specific data source
+    by its identifier. This class acts as a central repository for managing data sources in
+    bioinformatics applications. Currently DNA methylation data from GEO is supported.
+
+    Attributes:
+        sources (list): A list of data sources loaded from the library file.
+    """
     def __init__(self, library_file=None):
+        """
+        Initializes the DataLibrary instance.
+
+        Args:
+            library_file (str, optional): The file path of the library file. If None, 
+                                          the biolearn default library will be loaded.
+
+        """
         self.sources = []
         if library_file is None:
             library_file = get_data_file("library.yaml")
         self.load_sources(library_file)
 
     def load_sources(self, library_file):
+        """
+        Loads data sources from a given library file appending them to the current set of data sources.
+
+        Args:
+            library_file (str): The file path of the library file to load data sources from.
+        """
         with open(library_file, "r") as f:
             data_sources = parse_library_file(f)
             self.sources.extend(data_sources)
 
     def get(self, source_id):
+        """
+        Retrieves a data source by its identifier.
+
+        Args:
+            source_id (str): The identifier of the data source to retrieve.
+
+        Returns:
+            The data source with the given identifier if found, otherwise None.
+        """
         for source in self.sources:
             if source.id == source_id:
                 return source
         return None
 
     def lookup_sources(self, organism=None, format=None):
+        """
+        Looks up data sources based on the specified organism and/or format.
+
+        Args:
+            organism (str, optional): The organism to filter the data sources by.
+            format (str, optional): The format to filter the data sources by.
+
+        Returns:
+            A list of data sources that match the specified organism and format criteria.
+        """
         matches = []
         for source in self.sources:
             if (organism is None or source.organism == organism) and (
@@ -157,3 +260,4 @@ class DataLibrary:
             ):
                 matches.append(source)
         return matches
+
