+++ doc/modules/data_library.rst
@@ -1,0 +1,19 @@
+.. _data_library_ref:
+
+:mod:`biolearn.data_library`: DataLibrary
+=================================
+
+.. automodule:: biolearn.data_library
+   :no-members:
+   :no-inherited-members:
+
+**Classes**:
+
+.. currentmodule:: biolearn.data_library
+
+.. autosummary::
+   :toctree: generated/
+   :template: class.rst
+
+   DataLibrary
+   DataSource
